requirements.txt
paho-mqtt==1.6.1
redis==4.6.0
